Reorganize components/containers
