+++ src/app/components/index.ts
@@ -1,0 +1,4 @@
+export * from './header/header.component';
+export * from './icon/icon.component';
+export * from './picker/picker.component';
+export * from './select/select.component';
