+++ src/app/components/select/select.component.html
@@ -1,0 +1,1 @@
+<p>select works!</p>
