+++ src/app/containers/index.ts
@@ -1,0 +1,3 @@
+export * from './products-container';
+export * from './settings-container';
+export * from './steps-container';
