+++ src/app/containers/products-container/index.ts
@@ -1,0 +1,2 @@
+export * from './products/products.component';
+export * from './products-container.component';
