+++ src/app/containers/settings-container/index.ts
@@ -1,0 +1,2 @@
+export * from './settings/settings.component';
+export * from './settings-container.component';
