+++ src/app/containers/settings-container/settings/settings.component.html
@@ -139,3 +139,7 @@
     <label for="settings-precision-factory-fractions">Fractions</label>
   </div>
 </div>
+<div class="label">Assembler</div>
+<div class="field"></div>
+<div class="label">Furnace</div>
+<div class="field"></div>
