+++ src/app/containers/steps-container/index.ts
@@ -1,0 +1,2 @@
+export * from './steps/steps.component';
+export * from './steps-container.component';
