+++ src/app/containers/steps-container/steps/steps.component.scss
